requirements.txt
pandas==2.2.2
mysql-connector-python==9.0.0
pymongo==4.8.0
redis==5.0.7
scikit-learn==1.5.1
joblib==1.4.2
shap==0.46.0
numpy==2.0.0
matplotlib==3.9.1
streamlit==1.36.0